requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
pandas>=2.0.0
pillow>=10.0.0
scikit-image>=0.22.0
tqdm>=4.66.0
matplotlib>=3.7.0
nltk>=3.8.1
rouge-score>=0.1.2
pydicom>=2.3.0
gradio>=4.0.0
